Set game instance to our custom game instance in project settings.
PS. Because of changes in file structure, note that existing project settings may be reset to default. I already put the main menu as the editor start up and game launch level. Also, this process broke the save game for settings... fixed by deleting it and just letting the code recreate it. Odd