Imported custom icons for objective marker.
Added an enumerator that can be set in WBP_ObjectiveMarker that allows for selecting which icon you want to display.

Added icons for different objectives into objective display. Also made it so that the destruction meter objective only displays if the power plants have been destroyed before the destruction meter is full.

Made the default display for objective marker the bonus objective

Added Objective marker component to BP_Lavatunnel for when it rises, the marker will appear.

Added Objective marker component to boss spawner that appears until the boss is hit by the player, in which case it goes away. (On defeat would require some BPI to create a signal for the spawner to react to. The bosses don't universally get destroyed or send a signal to react to)

Need to add bonus objective marker implementation for the power plants as well, but I don't understand them well enough to.