Fixing issues before build
The level select now sends the player to Mamba_Metro_V2, as the original is a week out of date.
The lava tunnel is in all 3 levels, and of the appropriate scale.
The Lava tunnel triggers to appear now when the level is completable.
There is a boss spawn with Birthday in both Hawiwi City and in Mamba Metropolis.
And lastly, the SFX in the settings menu have a minimum time between playing back to back. Minimizing spam.
The Async loading screen is not on because in branches, somebody disabled the plugin, then pushed that disable which also blows up all my set up in it. If we want to keep it, I'll have to do rework on that later on